refactor:
- Update imports in FunctionEditor and BlockAdder to reflect the new folder structure
